# bthwani-backend-nest/reports/openapi.yaml
openapi: 3.0.0
paths:
...
  /notifications:
    post:
      operationId: NotificationController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNotificationDto'
      responses:
        '201':
          description: ''
      security:
        - bearer: []
      summary: إنشاء إشعار (للإدارة)
      tags:
        - Notification
  /notifications/my:
    get:
      operationId: NotificationController_getMyNotifications
      parameters:
        - name: cursor
          required: false
          in: query
          description: Cursor للصفحة التالية
          schema:
            type: string
        - name: limit
          required: false
          in: query
          description: عدد العناصر
          schema:
            default: 20
            type: number
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      summary: جلب إشعارات المستخدم
      tags:
        - Notification
  /notifications/{id}/read:
    post:
      operationId: NotificationController_markAsRead
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: ''
      security:
        - bearer: []
      summary: تحديد الإشعار كمقروء
      tags:
        - Notification
  /notifications/{id}:
    delete:
      operationId: NotificationController_delete
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      summary: حذف إشعار
      tags:
        - Notification
  /notifications/suppression:
    post:
      operationId: NotificationController_createSuppression
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSuppressionDto'
      responses:
        '201':
          description: ''
      security:
        - bearer: []
      summary: حظر قنوات إشعارات محددة
      tags:
        - Notification
    get:
      operationId: NotificationController_getUserSuppressions
      parameters: []
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      summary: جلب قائمة الحظر للمستخدم
      tags:
        - Notification
  /notifications/suppression/channels:
    get:
      operationId: NotificationController_getSuppressedChannels
      parameters: []
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      summary: جلب القنوات المحظورة
      tags:
        - Notification
  /notifications/suppression/{id}:
    delete:
      operationId: NotificationController_removeSuppression
      parameters:
        - name: id
          required: true
          in: path
          description: معرف الحظر
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      summary: إلغاء حظر
      tags:
        - Notification
  /notifications/suppression/channel/{channel}:
    delete:
      operationId: NotificationController_removeChannelSuppression
      parameters:
        - name: channel
          required: true
          in: path
          schema:
            enum:
              - push
              - email
              - sms
            type: string
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      summary: إلغاء حظر قناة محددة
      tags:
        - Notification
  /notifications/suppression/stats:
    get:
      operationId: NotificationController_getSuppressionStats
      parameters: []
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      summary: إحصائيات الحظر (للإدارة)
      tags:
        - Notification
...
  /legal/privacy-policy:
    get:
      operationId: LegalController_getPrivacyPolicy
      parameters:
        - name: lang
          required: false
          in: query
          description: اللغة (ar أو en)
          schema:
            enum:
              - ar
              - en
            type: string
      responses:
        '200':
          description: تم الحصول على سياسة الخصوصية بنجاح
        '404':
          description: لا توجد سياسة خصوصية نشطة
      summary: الحصول على سياسة الخصوصية النشطة
      tags:
        - Legal
  /legal/terms-of-service:
    get:
      operationId: LegalController_getTermsOfService
      parameters:
        - name: lang
          required: false
          in: query
          description: اللغة (ar أو en)
          schema:
            enum:
              - ar
              - en
            type: string
      responses:
        '200':
          description: تم الحصول على شروط الخدمة بنجاح
        '404':
          description: لا توجد شروط خدمة نشطة
      summary: الحصول على شروط الخدمة النشطة
      tags:
        - Legal
  /legal/consent:
    post:
      operationId: LegalController_recordConsent
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordConsentDto'
      responses:
        '201':
          description: تم تسجيل الموافقة بنجاح
        '400':
          description: بيانات غير صحيحة
        '401':
          description: غير مصرح
      security:
        - bearer: []
      summary: تسجيل موافقة المستخدم
      tags:
        - Legal
  /legal/consent/my:
    get:
      operationId: LegalController_getMyConsents
      parameters: []
      responses:
        '200':
          description: تم الحصول على الموافقات بنجاح
        '401':
          description: غير مصرح
      security:
        - bearer: []
      summary: الحصول على موافقات المستخدم الحالي
      tags:
        - Legal
  /legal/consent/check/{type}:
    get:
      operationId: LegalController_checkConsent
      parameters:
        - name: type
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: تم التحقق بنجاح
        '401':
          description: غير مصرح
      security:
        - bearer: []
      summary: التحقق من موافقة المستخدم على الإصدار الحالي
      tags:
        - Legal
  /legal/admin/privacy-policies:
    get:
      operationId: LegalController_getAllPrivacyPolicies
      parameters: []
      responses:
        '200':
          description: تم الحصول على السياسات بنجاح
        '401':
          description: غير مصرح
        '403':
          description: ممنوع - يتطلب صلاحيات إدارية
      security:
        - bearer: []
      summary: الحصول على جميع سياسات الخصوصية (للإدارة)
      tags:
        - Legal
  /legal/admin/privacy-policy:
    post:
      operationId: LegalController_createPrivacyPolicy
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePrivacyPolicyDto'
      responses:
        '201':
          description: تم إنشاء السياسة بنجاح
        '400':
          description: بيانات غير صحيحة
        '401':
          description: غير مصرح
        '403':
          description: ممنوع - يتطلب صلاحيات إدارية
      security:
        - bearer: []
      summary: إنشاء سياسة خصوصية جديدة (للإدارة)
      tags:
        - Legal
  /legal/admin/privacy-policy/{id}/activate:
    patch:
      operationId: LegalController_activatePrivacyPolicy
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: تم تفعيل السياسة بنجاح
        '401':
          description: غير مصرح
        '403':
          description: ممنوع - يتطلب صلاحيات إدارية
        '404':
          description: السياسة غير موجودة
      security:
        - bearer: []
      summary: تفعيل سياسة خصوصية معينة (للإدارة)
      tags:
        - Legal
  /legal/admin/terms-of-service:
    get:
      operationId: LegalController_getAllTermsOfService
      parameters: []
      responses:
        '200':
          description: تم الحصول على الشروط بنجاح
        '401':
          description: غير مصرح
        '403':
          description: ممنوع - يتطلب صلاحيات إدارية
      security:
        - bearer: []
      summary: الحصول على جميع شروط الخدمة (للإدارة)
      tags:
        - Legal
    post:
      operationId: LegalController_createTermsOfService
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTermsOfServiceDto'
      responses:
        '201':
          description: تم إنشاء الشروط بنجاح
        '400':
          description: بيانات غير صحيحة
        '401':
          description: غير مصرح
        '403':
          description: ممنوع - يتطلب صلاحيات إدارية
      security:
        - bearer: []
      summary: إنشاء شروط خدمة جديدة (للإدارة)
      tags:
        - Legal
  /legal/admin/terms-of-service/{id}/activate:
    patch:
      operationId: LegalController_activateTermsOfService
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: تم تفعيل الشروط بنجاح
        '401':
          description: غير مصرح
        '403':
          description: ممنوع - يتطلب صلاحيات إدارية
        '404':
          description: الشروط غير موجودة
      security:
        - bearer: []
      summary: تفعيل شروط خدمة معينة (للإدارة)
      tags:
        - Legal
  /legal/admin/consent/statistics:
    get:
      operationId: LegalController_getConsentStatistics
      parameters: []
      responses:
        '200':
          description: تم الحصول على الإحصائيات بنجاح
        '401':
          description: غير مصرح
        '403':
          description: ممنوع - يتطلب صلاحيات إدارية
      security:
        - bearer: []
      summary: الحصول على إحصائيات الموافقات (للإدارة)
      tags:
        - Legal
...
info:
  title: Bthwani API v2
  description: NestJS API Documentation - نظام إدارة الطلبات والتجارة
  version: '2.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    CreateNotificationDto:
      type: object
      properties:
        toUser:
          type: string
          description: معرف المستخدم المستهدف
        audience:
          description: الجمهور المستهدف
          type: array
          items:
            type: string
        title:
          type: string
          description: العنوان
        body:
          type: string
          description: المحتوى
        data:
          type: object
          description: بيانات إضافية
        collapseId:
          type: string
          description: معرف التجميع
      required:
        - title
        - body
    CreateSuppressionDto:
      type: object
      properties:
        suppressedChannels:
          type: array
          description: القنوات المراد حظرها
          example:
            - email
            - push
          items:
            type: string
            enum:
              - push
              - email
              - sms
        reason:
          type: string
          description: سبب الحظر
          enum:
            - user_request
            - bounce
            - complaint
            - unsubscribe
            - invalid_contact
            - too_many_failures
            - admin_block
          example: user_request
        details:
          type: string
          description: تفاصيل إضافية
          example: المستخدم طلب إيقاف الإشعارات
        expiresAt:
          format: date-time
          type: string
          description: تاريخ انتهاء الحظر
          example: '2025-12-31T23:59:59Z'
      required:
        - suppressedChannels
        - reason
...
    RecordConsentDto:
      type: object
      properties:
        consentType:
          type: string
          description: نوع الموافقة
          enum:
            - privacy_policy
            - terms_of_service
          example: privacy_policy
        version:
          type: string
          description: إصدار المستند
          example: 1.0.0
        accepted:
          type: boolean
          description: هل تمت الموافقة
          default: true
        ipAddress:
          type: string
          description: عنوان IP
        userAgent:
          type: string
          description: User Agent
      required:
        - consentType
        - version
        - accepted
    CreatePrivacyPolicyDto:
      type: object
      properties:
        version:
          type: string
          description: إصدار السياسة
          example: 1.0.0
        content:
          type: string
          description: محتوى السياسة بالعربية
        contentEn:
          type: string
          description: محتوى السياسة بالإنجليزية
        effectiveDate:
          format: date-time
          type: string
          description: تاريخ سريان السياسة
        isActive:
          type: boolean
          description: هل السياسة نشطة
          default: true
      required:
        - version
        - content
        - contentEn
    CreateTermsOfServiceDto:
      type: object
      properties:
        version:
          type: string
          description: إصدار الشروط
          example: 1.0.0
        content:
          type: string
          description: محتوى الشروط بالعربية
        contentEn:
          type: string
          description: محتوى الشروط بالإنجليزية
        effectiveDate:
          format: date-time
          type: string
          description: تاريخ سريان الشروط
        isActive:
          type: boolean
          description: هل الشروط نشطة
          default: true
      required:
        - version
        - content
        - contentEn